Unenforced DHCP
Before, You had to assign a machine to an IP part in the lab JSON file (For example, ["Machine3", "x.x.x.4"]) inside the networks' configuration  section. This is no longer required and if a machine is assigned to a network and is not assigned an IP address then the DHCP server will assign it automatically. This further reduces required configuration and reduces possible misconfiguration when designing labs.